ASHA/requirements.txt
streamlit==1.44.1
google-generativeai==0.8.5
python-dotenv==1.1.0
pillow==11.2.1 